added difficulties in battle

See merge request !157

+++ src/battle/view/EnemyView.java
@@ -1,5 +1,6 @@
 package battle.view;
 
+
 /**
  * Created by Bill Xiong on 12/7/16.
  * 
@@ -7,9 +8,7 @@ package battle.view;
  */
 public class EnemyView extends ItemView {
 	private static final String ENEMY_NAME = "Enemy";
-
-	public EnemyView(int hp, int x, int y, String filePath) {
+	public EnemyView(double hp, int x, int y, String filePath) {
 		super(ENEMY_NAME, hp, x, y,filePath);
-		
 	}
 }
